Update 4
Added bodyparser middleware to parse body text in a post response into a an object on the request object.  Created sendecho server route that returns json with a route parameter.  Created a name route that returns a json object containing the post query parameters.  Began work on a send/response route for form post and response.