url-short-update-01
Added skeleton project.  Began work creating post callback and route.  App takes a url and responds with json containing the url and a new shortened route to visit.  Visiting the route redirects the user to the original posted url.  Added body parser middleware to parse the posted url and save for use.